rename the test to remove redundancy
